Ajustando as classas Cliente e Endereco

+++ src/main/java/com/app_pedidos/model/services/ProdutoService.java
@@ -12,7 +12,7 @@ public class ProdutoService {
     @Autowired
     private ProdutoRepository repository;
 
-    public void inativar(int id) {
+    public void inativar(Long id) {
         Produto produtoExcluir = repository.getById(id);
         produtoExcluir.setStatus(false);
         repository.save(produtoExcluir);
